Add support for tmux windows
If a long running command is run from a tmux window, on completion, it
verifies that the tmux window differs before triggering a notification.

Information about the tmux session related to the completed
command is added to the notification to make it easier to figure out in
which tmux window the command was being executed.